updated academy pages
