--- src/data.js
@@ -1,62 +1,0 @@
-import programImage from "./images/academy/cloud-program.svg";
-import onlineCourseImage from "./images/academy/online-course.svg";
-
-export const programsList = [
-  {
-    id: 0,
-    title: "Cloud Programs",
-    description:
-      "Our Cloud Programs offer a variety of learning initiatives And bootcamps targeted at getting more African women into the Cloud Engineering Space, starting with our cohortly 3-months Cloud School bootcamp",
-    image: programImage,
-  },
-  {
-    id: 1,
-    title: "Cloud Programs",
-    description:
-      "Our Cloud Programs offer a variety of learning initiatives And bootcamps targeted at getting more African women into the Cloud Engineering Space, starting with our cohortly 3-months Cloud School bootcamp",
-    image: programImage,
-  },
-];
-export const onlineCoursesList = [
-  {
-    id: 0,
-    description:
-      "Master of Molecular Science & Software Engineering Online Degree Program",
-    image: onlineCourseImage,
-  },
-  {
-    id: 1,
-    description:
-      "Master of Molecular Science & Software Engineering Online Degree Program",
-    image: onlineCourseImage,
-  },
-  {
-    id: 2,
-    description:
-      "Master of Molecular Science & Software Engineering Online Degree Program",
-    image: onlineCourseImage,
-  },
-  {
-    id: 3,
-    description:
-      "Master of Molecular Science & Software Engineering Online Degree Program",
-    image: onlineCourseImage,
-  },
-];
-export const youtubeVideos = [
-  {
-    id: 0,
-    description: "Nicole's story - Chasing your dreams in Tech",
-    video: "https://www.youtube.com/embed/wBVYJWAH2B0",
-  },
-  {
-    id: 1,
-    description: "Ayodele's Story",
-    video: "https://www.youtube.com/embed/uwpBqSkJTPg",
-  },
-  {
-    id: 2,
-    description: "Bami's Story",
-    video: "https://www.youtube.com/embed/USWzE09j0M8",
-  },
-];
