+++ src/utils/index.js
@@ -4,6 +4,9 @@ import layer5 from "../images/Layer-5.png";
 import chimoney from "../images/chimoney.png";
 import deimos from "../images/deimos.png";
 import cloudbees from "../images/cloudbees.png";
+import programImage from "../images/academy/cloud-program.svg";
+import onlineCourseImage from "../images/academy/online-course.svg";
+
 
 export const paths = {
 home: "/",
@@ -12,7 +15,7 @@ home: "/",
   engineering: "/school-of-engineering",
   products: "/school-of-products",
   appliedSkills: "/school-of-applied-skills",
-  stem: "/stem-school",
+  stemSchool: "/stem-school",
   innitiatives: "/innitiatives",
   chapters: "/chapters",
   donate: "/donate-partner",
@@ -35,6 +38,7 @@ home: "/",
   faq: "/faq",
   codeOfConduct: "/code-of-conduct",
   privacy: "/privacy",
+  
 };
 
 export const apiConstants = {
@@ -703,5 +707,68 @@ export const whoWeAreTexts = [
   {
     heading: "our core values",
     text: "Team work, Community, Technical growth, Leadership, Visibility.",
+  },
+];
+
+
+
+// Accademy pages
+export const programsList = [
+  {
+    id: 0,
+    title: "Cloud Programs",
+    description:
+      "Our Cloud Programs offer a variety of learning initiatives And bootcamps targeted at getting more African women into the Cloud Engineering Space, starting with our cohortly 3-months Cloud School bootcamp",
+    image: programImage,
+  },
+  {
+    id: 1,
+    title: "Cloud Programs",
+    description:
+      "Our Cloud Programs offer a variety of learning initiatives And bootcamps targeted at getting more African women into the Cloud Engineering Space, starting with our cohortly 3-months Cloud School bootcamp",
+    image: programImage,
+  },
+];
+export const onlineCoursesList = [
+  {
+    id: 0,
+    description:
+      "Master of Molecular Science & Software Engineering Online Degree Program",
+    image: onlineCourseImage,
+  },
+  {
+    id: 1,
+    description:
+      "Master of Molecular Science & Software Engineering Online Degree Program",
+    image: onlineCourseImage,
+  },
+  {
+    id: 2,
+    description:
+      "Master of Molecular Science & Software Engineering Online Degree Program",
+    image: onlineCourseImage,
+  },
+  {
+    id: 3,
+    description:
+      "Master of Molecular Science & Software Engineering Online Degree Program",
+    image: onlineCourseImage,
+  },
+];
+export const youtubeVideos = [
+  {
+    id: 0,
+    description: "Nicole's story - Chasing your dreams in Tech",
+    video: "https://www.youtube.com/embed/wBVYJWAH2B0",
+  },
+  {
+    id: 1,
+    description: "Ayodele's Story",
+    video: "https://www.youtube.com/embed/uwpBqSkJTPg",
+  },
+  {
+    id: 2,
+    description: "Bami's Story",
+    video: "https://www.youtube.com/embed/USWzE09j0M8",
   },
 ];
